compat(win8): decimal quantities, egg-tray fractional sales, piece defaults
Second wave of the Fahd rollout fixes, ported to Dart 3.3:
- purchasing quantities are decimals end-to-end (draft, receive,
  adjustments): fractional units (half an egg tray, 2.5 kg) transact in
  fractions; whole-number units keep rejecting them; quantities go on
  the wire as 3dp strings matching the backend
- tap-to-type quantity entry on the draft line stepper, decimal
  keyboards on receive/adjustment fields, quantity displays formatted
  without trailing zeros
- POS cart lines of multi-unit products open the unit+quantity sheet
  from the quantity tap (selling half an egg tray at the till)
- l10n quantity placeholders take formatted strings; regenerated

Dart 3.3 ports: null-aware collection element in the PO payload and a
duplicate wildcard param rewritten to pre-3.7 syntax.

+++ frontend/lib/src/features/purchasing/views/purchase_catalog_pane.dart
@@ -171,7 +171,7 @@ class PurchaseCatalogPane extends StatelessWidget {
     for (final entry in entries) {
       await viewModel.addVariant(
         entry.variant,
-        quantity: entry.quantity,
+        quantity: entry.quantity.toDouble(),
         source: 'purchase_camera_scanner',
       );
     }
